Fix bug in findNode() (+)
(+) findNode was fixed before so that it could search through a
poseCommandGroup object. However, this is a slightly different
structure than normal commandGroup objects, so when the children of the
poseCommandGroup (which are commandGroups) were passed into the
function, it would throw an error and fail to find anything inside of
the children of the poseCommandGroup.

Now, it has an extra argument that tells whether or not this is an initial
call of the function, and searches differently based upon the node. In
hindsight, this is not the most efficient way to fix that error, and I
will be fixing the way I fixed it soon.

pain
- Emmie