Remove options from Pose
The constructor function of `Pose` was given an final parameter of
`options` to act as a container for additional arguments that the
constructor would need.

The benefit of this `options` parameter, is it keeps the parameter count
of `Pose` low, and we can prepare an object with all the arguments we
need prior to calling `Pose`, which makes the code a bit easier to read
than 6+ arguments that might require default value checks.

We'll keep the `options` parameter.  It's able to do it's job because it
holds and object value.  But once we are inside `Pose`, we begin the
process of creating a new object to be a `Pose` value.  We don't need to
keep all the arguments inside `options`, but can instead spread them out
in the new `Pose` object we're creating.

This reduces all the code that uses `pose.options.commands` down to
`pose.commands`, which is more succinct.  When we're working with a
pose, we don't care to remember how we passed extra arguments to its
constructor function, and `pose.options` is a constant reminder.

- Justin C