Remove drawing context from makeRotation()
Be wary of drawing at arbitrary moments of time in the execution of the
program.  It can lead to glitches and artifacts you will sink time into
debugging.

makeRotation() has one job, make a new rotation object.  Well, it also
has a second job: add the new rotation object to the list.

There is one moment in time that the program draws everything, and that
happens in _redrawCanvas(), which already draws rotations.

After creating a rotation object, if the screen needs redrawn to show
the new rotation, then call redrawCanvas() to schedule _redrawCanvas().

- Justin C