Delete rotPts no longer on valid path+0, !
(+0) - Deletes rotations that were on an end path that was deleted or
otherwise could not be placed onto a valid t value along the path. Also
re-adds the initial rotation point if all rotations are removed.
However, see the bug here below:

(!) - Only works if all rotations are removed; doesn't factor in the
fact that the first waypoint can be deleted without all rotations going
with it. Also doesn't factor in if there is only one waypoint. Will fix.

- Emmie Jones (War Criminal)