Move CmdPts via select tool
Have to fully grab the point itself. This is to make it easier to
distinguish when grabbing a cmdPt on a pose. Radius of 10px assumed
enough space. If this is inaccessible, it'll probably be better to make
a separate accessibility mode for that. (Although current accessibility
is pretty good.)

Bravi! The Genetic Opera Is Good!
- Emmie Jones