Refactor bakeAdvancedExport to accept command list argument
An effort was made previously to make bakeAdvancedExport as a pure
function.  All it's execution relies only on the arguments given, and
nothing else outside the function.

Technically, it calls impure functions that reach out to global
variables like poseList, but we should strive toward pure functions.

Pure functions are easier to test with automated testing (if/when we
decide to), can be easier to reason/understand, and makes it easier to
reuse the functions should we find ourselves with more than one command
list.

So, with that in mind, let's pass in the command points as an argument
like we do with the rotation points.

- Justin C